+++ packages/json-rpc/ws-provider/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.2.0 - 2024-08-12
+
 ### Changed
 
 - Renamed `WebSocketProvider` => `getWsProvider`